Delete README.md Branch
Secretly changing or inserting name information is to put the blame on me. I didn't think to name the company exactly the same.  But there must have been an original intention (because of the correct line offering) to bring programs that were no longer supported or were no longer used.  To be used in the basic education of underprivileged people or remote communities who may not have the money to buy the program or even if the price is too high if you choose food with the program while your stomach is hungry.  What will you choose?  And why? If you don't use it to attack me again many times and sneak into my app many times and still use my personal information, you will bring the agency that  So, are you coming to pay me back? What do you say? At this point, none of the people who have come to this point can read or are not good at translating. But the translation has improved to sixty-five percent, but they can communicate.  I can understand it a bit, but you guys like to claim that it's this or that case because of your soft intelligence.  Some coincidences, etc. In short, it is hacking, duplicating, modifying data in other people's personal space. What else do you want?  Either way, I won't cheat.  Don't rob anyone to eat.  I don't create division among anyone, and if you do something wrong, focus on your own fault.  Do not bring up other people or their faults to avoid your point or you will get one more charge.patch